Update RadioButton component with new prop
This commit introduces several updates to the RadioButton component. New optional props `error` and `disabled` have been added, and the `RadioButtonOptionType` has been updated with additional fields: `errorMessage`, `errorIcon`, `errorIconAltText`, `errorAriaLiveType`, `error`, and `disabled`. The existing `state` prop will be deprecated, and the new `checked`, `disabled`, and `error` props will be used to control the state of the RadioButton internally. These changes enhance the flexibility and usability of the RadioButton component.