Fix font-weight and color props of Link Component
This commit resolves two issues with the Link component where the font-weight and color props were not working as expected. Changes made include:

In link.tsx, added receipt of weight value via props. This was necessary as the styles.container object was not storing the weight value, resulting in it always being undefined.
In link.styled.ts, included the applyDevicePropsTextStyles function to the link's style. This was required as color is a text property and needs to be applied by the TextStylesExtended styles constant.
These changes ensure that changes to the font-weight and color props in Storybook are accurately reflected in the Link component's visual representation