Fix Font-Weight and Color Props of Link Component
This commit addresses two issues with the Link component: the 'font-weight' and 'color' props were not functioning correctly. When a user set a different value from the default in Storybook, the changes were not reflected visually in the component.

Key changes include:

link.tsx: Added receipt of 'weight' value via props. Previously, the 'styles.container' object wasn't storing the 'weight' value, so it was always undefined.
link.styled.ts: Included 'applyDevicePropsTextStyles' function to link's style. The 'color' is a text property and needs to be applied by 'TextStylesExtended' styles const.
These changes ensure that the 'font-weight' and 'color' props of the Link component function as expected, improving the component's flexibility and usability.